% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{cli_helpers}
\alias{cli_helpers}
\alias{bracketize}
\alias{stylize}
\alias{bulletize}
\title{Assorted helper functions to format text and vectors for cli messages}
\usage{
bracketize(..., .open = "{", .close = "}", collapse = NULL)

stylize(
  x,
  style = NULL,
  bracket = FALSE,
  .open = "{",
  .close = "}",
  collapse = NULL
)

bulletize(
  x,
  bullet = "*",
  sep = NULL,
  before = NULL,
  after = NULL,
  style = NULL,
  n_show = 5,
  n_fudge = 2,
  bracket = FALSE
)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{.open, .close}{Open and close bracket characters.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.  When \code{collapse} is a string,
    the result is always a string (\code{\link[base]{character}} of length 1).}

\item{x}{A vector.}

\item{style}{A cli style name, e.g. code, val, file, url}

\item{bracket}{If \code{TRUE}, pass x to bracketize.}

\item{bullet}{Character to use for bullet. Defaults to "".}

\item{sep, before, after}{Additional characters or character vectors applied
using \code{paste0(before, sep, out, after)}.}

\item{n_show}{The maximum number of items to include in the bullet list.
Defaults to 5.}

\item{n_fudge}{The minimum number of items to include in summary of additional
bullet items. If the summary would only include a number of items equal or
less than n_fudge, they are included in the bullet list and the summary is
not displayed. Defaults to 2.}
}
\description{
These functions are convenient alternatives to the internal cli functions but
may be removed as my understanding of the cli package improves.
\itemize{
\item \code{\link[=bracketize]{bracketize()}} pastes open and close brackets around a vector.
\item \code{\link[=stylize]{stylize()}} appends a style prefix, e.g. "\verb{\{.val value\}}" to a vector.
\item \code{\link[=bulletize]{bulletize()}} turns a vector into a neatly abbreviated bullet list.
}

\code{\link[=bulletize]{bulletize()}} is adapted from the gargle package (see
\href{https://github.com/r-lib/gargle/blob/4021167fd2f7aca7194027bf73a5a06296ca03dc/R/utils-ui.R}{\code{utils-ui.R}}).
This function is available under a MIT license and is the work of the gargle
authors.
}
\examples{
bracketize("value")

stylize("styled value", "val")

stylize("styled variable", "val", bracket = TRUE)

bulletize(c("val 1", "val 2", "val 3"))

bulletize(rep("val", 20), n_show = 3)

}
