% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_paths.R
\name{cli_paths}
\alias{cli_paths}
\title{Display a list of file paths}
\usage{
cli_paths(path, msg)
}
\arguments{
\item{path}{One or more file paths.}

\item{msg}{Passed to message for \code{\link[cli:cli_inform]{cli_inform()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
Use \code{\link[=cli_list_files]{cli_list_files()}} instead.
}
\seealso{
\code{\link[cli:cli_bullets]{cli::cli_bullets()}}
}
