% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExtras.R
\name{cli_ul_items}
\alias{cli_ul_items}
\title{Format a list of items as an unordered list with cli_ul()}
\usage{
cli_ul_items(items)
}
\description{
Format a list of items as an unordered list with cli_ul()
}
