% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExtras-package.R
\docType{package}
\name{cliExtras-package}
\alias{cliExtras}
\alias{cliExtras-package}
\title{cliExtras: A Collection of Helper Functions for the cli package}
\description{
Extends the cli package with user response prompts and stopifnot equivalents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/cliExtras}
  \item \url{https://elipousson.github.io/cliExtras/}
  \item Report bugs at \url{https://github.com/elipousson/cliExtras/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]
}

}
\keyword{internal}
