% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_cli_format.R
\name{register_cli_format}
\alias{register_cli_format}
\title{Register a custom cli format method with registerS3method}
\usage{
register_cli_format(class = NULL, method = NULL, envir = parent.frame())
}
\arguments{
\item{class, method}{Object class and format function passed to
\code{\link[=registerS3method]{registerS3method()}}.}

\item{envir}{Environment where the S3 method is registered. Defaults to
parent.frame().}
}
\description{
Register a custom cli format method with registerS3method
}
