% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_quiet.R
\name{set_cli_quiet}
\alias{set_cli_quiet}
\title{Set the cli.default_handler and cliExtras.quiet options}
\usage{
set_cli_quiet(quiet = FALSE, msg = !quiet)
}
\arguments{
\item{quiet}{If \code{TRUE} set the cli.default_handler option to
\code{suppressMessages} and cliExtras.quiet option to \code{TRUE}. Defaults to
\code{FALSE}.}

\item{msg}{If \code{TRUE}, \code{\link[=set_cli_quiet]{set_cli_quiet()}} displays a message confirming the
option changes. If \code{FALSE}, the function does not display a message.
Defaults to \code{!quiet}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
