% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_ask.R
\name{cli_ask}
\alias{cli_ask}
\title{Display a Message then Read a Line from the Terminal}
\usage{
cli_ask(prompt = "?", message = NULL, ..., .envir = parent.frame())
}
\arguments{
\item{prompt}{Characters to show as user prompt in console following
displayed message. A non-breaking space is always placed after the prompt
before passing to \code{\link[=readline]{readline()}}.}

\item{message}{Primary prompt to user passed to \code{\link[cli:cli_abort]{cli::cli_inform()}}.}

\item{...}{Passed to \code{\link[rlang:abort]{rlang::abort()}}, \code{\link[rlang:abort]{rlang::warn()}} or
\code{\link[rlang:abort]{rlang::inform()}}.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
Display a Message then Read a Line from the Terminal
}
\seealso{
\code{\link[=cli_yesno]{cli_yesno()}}
}
