% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExtras.R
\name{cli_ask}
\alias{cli_ask}
\title{Display a Message then Read a Line from the Terminal}
\usage{
cli_ask(message, ..., prompt = ">> ", .envir = parent.frame())
}
\arguments{
\item{message}{Primary prompt to user prefixed by ">" when passed to
\code{\link[=cli_inform]{cli_inform()}}.}

\item{...}{Secondary messages passed to \code{\link[=cli_inform]{cli_inform()}}}

\item{prompt}{Character to show as user prompt in console following displayed
message.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
Display a Message then Read a Line from the Terminal
}
