% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_ask.R
\name{cli_ask}
\alias{cli_ask}
\title{Display a Message then Read a Line from the Terminal}
\usage{
cli_ask(prompt = "?", ..., .envir = rlang::caller_env(), call = .envir)
}
\arguments{
\item{prompt}{Characters to show as user prompt in console following
displayed message. A non-breaking space is always placed after the prompt
before passing to \code{\link[=readline]{readline()}}. Defaults to "?".}

\item{...}{
  Arguments passed on to \code{\link[cli:cli_bullets]{cli::cli_bullets}}
  \describe{
    \item{\code{text}}{Character vector of items. See details below on how names
are interpreted.}
    \item{\code{id}}{Optional id of the \code{div.bullets} element, can be used in themes.}
    \item{\code{class}}{Optional additional class(es) for the \code{div.bullets} element.}
  }}

\item{.envir}{Environment to evaluate the glue expressions in.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Display a Message then Read a Line from the Terminal
}
\seealso{
\code{\link[=cli_yesno]{cli_yesno()}}
}
