% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_bulletize.R
\name{cli_bulletize}
\alias{cli_bulletize}
\title{List of items using bulletize helper}
\usage{
cli_bulletize(
  items,
  bullet = "*",
  n_show = Inf,
  n_fudge = 2,
  style = NULL,
  sep = NULL,
  before = NULL,
  after = NULL,
  id = NULL,
  class = NULL,
  .envir = parent.frame()
)
}
\arguments{
\item{items}{A named vector or list to use in creating a bulletted list with
\code{\link[cli:cli_bullets]{cli::cli_bullets()}}.}

\item{bullet}{Character to use for bullet. Defaults to "".}

\item{n_show}{The maximum number of items to include in the bullet list.
Defaults to 5.}

\item{n_fudge}{The minimum number of items to include in summary of additional
bullet items. If the summary would only include a number of items equal or
less than n_fudge, they are included in the bullet list and the summary is
not displayed. Defaults to 2.}

\item{style}{A cli style name, e.g. code, val, file, url}

\item{sep, before, after}{Additional characters or character vectors applied
using \code{paste0(before, sep, out, after)}. Defaults to \code{NULL}.}

\item{id}{Optional id of the \code{div.bullets} element, can be used in themes.}

\item{class}{Optional additional class(es) for the \code{div.bullets} element.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
List of items using bulletize helper
}
