% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExtras.R
\name{cli_menu}
\alias{cli_menu}
\title{Display a Message then Read a Line from the Terminal}
\usage{
cli_menu(
  choices,
  title = NULL,
  message = NULL,
  prompt = ">> ",
  choice_style = c("num", "val"),
  ...
)
}
\arguments{
\item{title}{title for menu}

\item{message}{Primary prompt to user passed to \code{\link[=cli_inform]{cli_inform()}}.}

\item{prompt}{Characters to show as user prompt in console following
displayed menu options.}

\item{...}{Secondary messages passed to \code{\link[=cli_inform]{cli_inform()}} with title as message.}
}
\description{
Display a Message then Read a Line from the Terminal
}
