% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_menu.R
\name{cli_menu}
\alias{cli_menu}
\title{Display a Message then Read a Line from the Terminal}
\usage{
cli_menu(
  choices,
  title = NULL,
  message = NULL,
  prompt = "Selection: ",
  style = c("num", "val"),
  labels = "arabic",
  sep = ". ",
  bullet = "",
  call = .envir,
  .envir = parent.frame()
)
}
\arguments{
\item{choices}{Required list of choices. If named, use the names as the
choice options.}

\item{title}{Title for menu. Passed to \code{\link[cli:cli_h1]{cli::cli_h1()}}}

\item{message}{Additional message to display after choices and before prompt.}

\item{prompt}{Characters to show as user prompt in console following
displayed menu options. Defaults to "Selection: " (matching \code{\link[utils:menu]{utils::menu()}})}

\item{style}{A cli style name, e.g. code, val, file, url}

\item{labels}{Label style. Options include "arabic", "alph", "Alph", "alpha",
"Alpha", "roman", or "Roman".}

\item{bullet}{Character to use for bullet. Defaults to "".}

\item{call}{Default: \code{\link[=parent.frame]{parent.frame()}}. Passed to input checking functions
to improve error message traceback.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
Display a Message then Read a Line from the Terminal
}
\examples{
\dontrun{
if (interactive()) {
  cli_menu(c("A", "B", "C"))

  cli_menu(c("A", "B", "C"), labels = "Alpha")
}
}
}
