% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExtras.R
\name{cli_paths}
\alias{cli_paths}
\title{Display a list of file paths}
\usage{
cli_paths(path, ..., .envir = parent.frame())
}
\arguments{
\item{path}{One or more file paths.}

\item{...}{Passed to message for \code{\link[=cli_inform]{cli_inform()}}}

\item{.envir}{Environment passed to \code{\link[=cli_inform]{cli_inform()}}, Default: parent.frame()}
}
\description{
Display a list of file paths
}
\seealso{
\code{\link[cli:cli_bullets]{cli::cli_bullets()}}
}
