% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExtras.R
\name{cli_paths}
\alias{cli_paths}
\title{Display a list of file paths}
\usage{
cli_paths(path, msg)
}
\arguments{
\item{path}{One or more file paths.}

\item{msg}{Passed to message for \code{\link[=cli_inform]{cli_inform()}}}
}
\description{
Display a list of file paths
}
\seealso{
\code{\link[cli:cli_bullets]{cli::cli_bullets()}}
}
