% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_progress_pipe.R
\name{cli_progress_pipe}
\alias{cli_progress_pipe}
\title{Simplified cli progress message for combining with a pipe}
\usage{
cli_progress_pipe(
  data,
  ...,
  current = TRUE,
  clear = FALSE,
  .auto_close = TRUE,
  .envir = current_env(),
  time = NULL
)
}
\arguments{
\item{data}{Input data. Reference w/ "{data}" if .envir is currrent_env().}

\item{...}{
  Arguments passed on to \code{\link[cli:cli_progress_bar]{cli::cli_progress_bar}}
  \describe{
    \item{\code{name}}{This is typically used as a label, and should be short,
at most 20 characters.}
    \item{\code{status}}{New status string of the progress bar, if not \code{NULL}.}
    \item{\code{type}}{Type of the progress bar. It is used to select a default
display if \code{format} is not specified. Currently supported types:
\itemize{
\item \code{iterator}: e.g. a for loop or a mapping function,
\item \code{tasks}: a (typically small) number of tasks,
\item \code{download}: download of one file,
\item \code{custom}: custom type, \code{format} must not be \code{NULL} for this type.
}}
    \item{\code{total}}{Total number of progress units, or \code{NA} if it is unknown.
\code{cli_progress_update()} can update the total number of units. This is
handy if you don't know the size of a download at the beginning, and
also in some other cases. If \code{format} is set to \code{NULL}, \code{format} (plus
\code{format_done} and \code{format_failed}) will be updated when you change
\code{total} from \code{NA} to a number. I.e. default format strings will be
updated, custom ones won't be.}
    \item{\code{format}}{Format string. It has to be specified for custom progress
bars, otherwise it is optional, and a default display is selected
based on the progress bat type and whether the number of total units
is known. Format strings may contain glue substitution, the support
pluralization and cli styling. See \link[cli]{progress-variables} for special
variables that you can use in the custom format.}
    \item{\code{format_done}}{Format string for successful termination. By default
the same as \code{format}.}
    \item{\code{format_failed}}{Format string for unsuccessful termination. By
default the same as \code{format}.}
    \item{\code{auto_terminate}}{Whether to terminate the progress bar if the
number of current units reaches the number of total units.}
    \item{\code{extra}}{Extra data to add to the progress bar. This can be
used in custom format strings for example. It should be a named list.
\code{cli_progress_update()} can update the extra data. Often you can get
away with referring to local variables in the format string, and
then you don't need to use this argument. Explicitly including these
constants or variables in \code{extra} can result in cleaner code. In
the rare cases when you need to refer to the same progress bar from
multiple functions, and you can them to \code{extra}.}
  }}

\item{current}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{clear}{Whether to remove the progress bar from the screen after it has
terminated. Defaults to \code{FALSE} (\code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} uses \code{TRUE}).}

\item{.auto_close}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{.envir}{The environment to use for auto-termination and for glue
substitution. It is also used to find and set the current progress bar.
Defaults to \code{\link[=current_env]{current_env()}} (\code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} uses
\code{\link[=parent.frame]{parent.frame()}})}

\item{time}{Passed to \code{\link[=Sys.sleep]{Sys.sleep()}} for optional pause after displaying
progress message. Defaults to \code{NULL}.}
}
\description{
A convenience function for creating progress messages that use piped data as
an input.
}
\examples{
df <- data.frame("letters" = LETTERS, "numbers" = c(1:26))

df |>
  cli_pipe_progress_msg("Data has {nrow(data)} rows and {ncol(data)} columns.") |>
  head(2)
}
