% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_quiet.R
\name{cli_quiet}
\alias{cli_quiet}
\alias{set_cli_quiet}
\title{Helper functions to get or set the cliExtras.quiet option}
\usage{
cli_quiet(quiet = FALSE)

set_cli_quiet(quiet = FALSE, msg = !quiet)
}
\arguments{
\item{quiet}{Option value to set or default value if option is not set.
Defaults to \code{FALSE}.}

\item{msg}{If \code{TRUE}, \code{\link[=set_cli_quiet]{set_cli_quiet()}} displays a message confirming the
option changes. If \code{FALSE}, the function does not display a message.}
}
\description{
\code{\link[=cli_quiet]{cli_quiet()}} gets the cliExtras.quiet option and \code{\link[=set_cli_quiet]{set_cli_quiet()}} sets the
option.
}
