% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_quiet.R
\name{cli_quiet}
\alias{cli_quiet}
\alias{set_cli_quiet}
\title{Helper functions to get or set the cli.quiet option}
\usage{
cli_quiet(quiet = FALSE)

set_cli_quiet(quiet = FALSE)
}
\arguments{
\item{quiet}{Option value to set or default value if option is not set.
Defaults to \code{FALSE}.}
}
\description{
\code{\link[=cli_quiet]{cli_quiet()}} gets the cli.quiet option and \code{\link[=set_cli_quiet]{set_cli_quiet()}} sets the
option.
}
