% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExtras.R
\name{cli_ul_items}
\alias{cli_ul_items}
\title{Format a list of items as an unordered list with cli_ul()}
\usage{
cli_ul_items(items, style = c("code", "val"), sep = ": ")
}
\arguments{
\item{items}{Character vector of items, or \code{NULL}.}

\item{style}{Length 2 character vector indicating cli style to use for names
of items and style to use for items.}

\item{sep}{Length 1 character vector with separator between item names and
names.}
}
\description{
Format a list of items as an unordered list with cli_ul()
}
