% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_vec.R
\name{cli_vec_last}
\alias{cli_vec_last}
\alias{cls_vec}
\title{Helper to add custom cli style to a vector with vec_last as a parameter}
\usage{
cli_vec_last(x, style = list(), vec_last = " or ")

cls_vec(x, vec_last = "or")
}
\arguments{
\item{x}{Vector that will be collapsed by cli.}

\item{style}{Style to apply to the vector. It is used as a theme on
a \code{span} element that is created for the vector. You can set \code{vec-sep}
and \code{vec-last} to modify the separator and the last separator.}

\item{vec_last}{Used as value for "vec-last" item in style object.}
}
\description{
Helper to add custom cli style to a vector with vec_last as a parameter
}
\keyword{internal}
