% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExtras.R
\name{cli_yesno}
\alias{cli_yesno}
\title{Yes No with Variable Responses using cli}
\usage{
cli_yesno(
  message,
  yes = c("Yes", "Definitely", "For sure", "Yup", "Yeah", "I agree", "Absolutely"),
  no = c("No way", "Not now", "Negative", "No", "Nope", "Absolutely not"),
  n_yes = 1,
  n_no = 1,
  call = .envir,
  .envir = parent.frame()
)
}
\arguments{
\item{message}{Passed to [cli_inform()]}

\item{yes, no}{Character strings with yes and no options.}

\item{n_yes, n_no}{Number of yes and no options to provide in user prompt.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
This is a modified version of [yesno::yesno()] and [usethis::ui_yeah()] to
work with [cli_inform()].
}
