% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_quiet.R
\name{quiet_cli_inform}
\alias{quiet_cli_inform}
\alias{quiet_cli_bullets}
\alias{quiet_cli_alert}
\title{Quiet versions of \code{\link[=cli_inform]{cli_inform()}}, \code{\link[=cli_bullets]{cli_bullets()}}, and \code{\link[=cli_alert]{cli_alert()}}}
\usage{
quiet_cli_inform(..., .envir = parent.frame())

quiet_cli_bullets(..., .envir = parent.frame())

quiet_cli_alert(..., .envir = parent.frame())
}
\arguments{
\item{...}{Additional parameters passed to \code{\link[=cli_inform]{cli_inform()}}, \code{\link[=cli_bullets]{cli_bullets()}},
or \code{\link[=cli_alert]{cli_alert()}}}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
The implementation of the cli.quiet option is based on the implementation of
a similar setting in googlesheets4, googledrive, and gargle.
}
